api/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
asyncpg==0.28.0
boto3==1.28.0
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.103.2
firebase-admin==6.0.0
fonttools==4.42.1
h11==0.14.0
httptools==0.6.0
httpx==0.25.0
idna==3.4
iniconfig==2.0.0
multipart==0.2.4
packaging==23.2
Pillow==10.0.0
pillow-heif==0.11.1
pluggy==1.3.0
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
pyjwt[crypto]==2.4.0
PyMuPDF==1.23.3
PyMuPDFb==1.23.3
pytest==7.4.0
pytest-asyncio==0.21.0
python-dotenv==1.0.0
PyYAML==6.0.1
sniffio==1.3.0
starlette==0.27.0
SQLAlchemy==2.0.18
tomli==2.0.1
typing_extensions==4.7.1
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3